// cube/types/Rotations.ts
export enum FundamentalRotations {
  U = "U",
  U_ = "U'",
  D = "D",
  D_ = "D'",
  L = "L",
  L_ = "L'",
  R = "R",
  R_ = "R'",
  F = "F",
  F_ = "F'",
  B = "B",
  B_ = "B'",
  M = "M",
  M_ = "M'",
  E = "E",
  E_ = "E'",
  S = "S",
  S_ = "S'",
}

export enum Rotations {
  // U
  U = "U",
  U_ = "U'",
  U2 = "U2",
  U2_ = "U2'",
  u = "u",
  u_ = "u'",
  u2 = "u2",
  u2_ = "u2'",
  // D
  D = "D",
  D_ = "D'",
  D2 = "D2",
  D2_ = "D2'",
  d = "d",
  d_ = "d'",
  d2 = "d2",
  d2_ = "d2'",
  // L
  L = "L",
  L_ = "L'",
  L2 = "L2",
  L2_ = "L2'",
  l = "l",
  l_ = "l'",
  l2 = "l2",
  l2_ = "l2'",
  // R
  R = "R",
  R_ = "R'",
  R2 = "R2",
  R2_ = "R2'",
  r = "r",
  r_ = "r'",
  r2 = "r2",
  r2_ = "r2'",
  // F
  F = "F",
  F_ = "F'",
  F2 = "F2",
  F2_ = "F2'",
  f = "f",
  f_ = "f'",
  f2 = "f2",
  f2_ = "f2'",
  // B
  B = "B",
  B_ = "B'",
  B2 = "B2",
  B2_ = "B2'",
  b = "b",
  b_ = "b'",
  b2 = "b2",
  b2_ = "b2'",
  // M
  M = "M",
  M_ = "M'",
  M2 = "M2",
  M2_ = "M2'",
  // E
  E = "E",
  E_ = "E'",
  E2 = "E2",
  E2_ = "E2'",
  // S
  S = "S",
  S_ = "S'",
  S2 = "S2",
  S2_ = "S2'",
  // x
  x = "x",
  x_ = "x'",
  x2 = "x2",
  x2_ = "x2'",
  // y
  y = "y",
  y_ = "y'",
  y2 = "y2",
  y2_ = "y2'",
  // z
  z = "z",
  z_ = "z'",
  z2 = "z2",
  z2_ = "z2'",
}

export const isRotation = (rotation: string): rotation is Rotations =>
  Object.values(Rotations).includes(rotation as Rotations);